refactoring server file to add namespaces
